Tests: Move Update from stable to master to update.go
- Fix issue on https://jenkins.cilium.io/job/Ginkgo-CI-Tests-Pipeline/1702/

```
/home/jenkins/workspace/Ginkgo-CI-Tests-Pipeline/src/github.com/cilium/cilium/test/k8sT/Nightly.go:478
Cannot curl app1-service
Expected
    <bool>: false
to be true
/home/jenkins/workspace/Ginkgo-CI-Tests-Pipeline/src/github.com/cilium/cilium/test/k8sT/Nightly.go:500
```
